+++ src/packages/picker/doc.en.ts
@@ -8,7 +8,6 @@ export const Internation = {
     chooseTime: '请选择时间',
     cascade: '多级联动',
     async: '异步获取',
-    custom: '自定义按钮',
     validTime: '请选择有效日期',
     chooseDate: '选择日期',
     always: '永远有效',
@@ -55,7 +54,6 @@ export const Internation = {
     chooseTime: '請選擇時間',
     cascade: '多級聯動',
     async: '異步獲取',
-    custom: '自定義按鈕',
     chooseDate: '選擇日期',
     always: '永遠有效',
     monday: '周一',
@@ -101,7 +99,6 @@ export const Internation = {
     chooseTime: 'Choose Time',
     cascade: 'Cascade',
     async: 'Async',
-    custom: 'Slots',
     validTime: 'Valid Time',
     chooseDate: 'Choose Date',
     always: 'Always',
